Refactor prime number generation and sieve algorithm

+++ Dimik OJ - দ্বিমিক অনলাইন জাজ/২ - জোড়-বিজোড় ২/C++/even-odd-2.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    long long testCases;
+    int testCases;
     cin >> testCases;
 
     while (testCases--) {
